+++ spring-core/src/main/java/org/springframework/core/convert/support/StringToSingletonCollectionConverter.java
@@ -1,4 +1,24 @@
+/*
+ * Copyright 2002-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.springframework.core.convert.support;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Set;
 
 import org.springframework.core.CollectionFactory;
 import org.springframework.core.convert.ConversionService;
@@ -6,10 +26,6 @@ import org.springframework.core.convert.TypeDescriptor;
 import org.springframework.core.convert.converter.ConditionalGenericConverter;
 import org.springframework.core.convert.converter.GenericConverter;
 import org.springframework.lang.Nullable;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Set;
 
 /**
  * Converts a String to a Singleton Collection.
@@ -46,6 +62,7 @@ public final class StringToSingletonCollectionConverter implements ConditionalGe
 		if (source == null) {
 			return null;
 		}
+
 		String string = (String) source;
 		String[] fields = {string};
 
@@ -57,7 +74,8 @@ public final class StringToSingletonCollectionConverter implements ConditionalGe
 			for (String field : fields) {
 				target.add(field.trim());
 			}
-		} else {
+		}
+		else {
 			for (String field : fields) {
 				Object targetElement = this.conversionService.convert(field.trim(), sourceType, elementDesc);
 				target.add(targetElement);
