+++ spring-core/src/test/java/org/springframework/core/convert/support/StringToCollectionConverterTests.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2002-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.core.convert.support;
+
+import java.util.Collection;
+import java.util.List;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+import org.springframework.core.convert.TypeDescriptor;
+
+public class StringToCollectionConverterTests {
+
+	@Test
+	void test() {
+		Collection<String> expected = List.of("one,two");
+
+		StringToSingletonCollectionConverter converter
+				= new StringToSingletonCollectionConverter(new GenericConversionService());
+
+		Collection<String> actual = (List<String>) converter.convert("one,two",
+				TypeDescriptor.valueOf(String.class),
+				TypeDescriptor.valueOf(List.class));
+
+		Assertions.assertThat(actual).isEqualTo(expected);
+	}
+
+}
